move build and setup config to build_config.py
